Added C++ and Python functions for creating motions and skeletons procedurally (i.e. without having to load a motion or skeleton file). To create a skeleton in Python (C++ is similar): sk = assets.createSkeletonDefinition("myskeleton") j = sk.createJoint("myjoint", None) j2 = sk.createJoint("anotherjoint", j)
To create a motion in Python (C++ is similar):
m = assets.createMotion("mymotion")
m.addChannel("base", "XPos")
m.addChannel("base", "YPos")
data = FloatVec()
data.append(1.4)
data.append(3.2)
m.addFrame(0, data)

cleaned up some warnings

+++ trunk/core/smartbody/sbgui/src/resourceViewer/TreeItemInfoWidget.cpp
@@ -721,7 +721,7 @@ void AnimationBlendInfoWidget::updateWidget()
 	blendInfoObject->clearAttributes();	
 	attrWindow->cleanUpWidgets();
 	std::vector<SrVec> parameters = animBlend->getParameters();
-	for (unsigned int i=0;i<animBlend->getNumMotions();i++)
+	for (unsigned int i=0;i< (size_t) animBlend->getNumMotions();i++)
 	{
 		std::string motionName = animBlend->getMotionName(i);
 		SrVec& para = parameters[i];
